requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
openvino
python-multipart 